+++ src/MJS_WordPlugin_src/WordAddIn1/GenerateHTMLButton.XMLBuildTocBody.cs
@@ -1,6 +1,4 @@
 ﻿using System.Collections.Generic;
-using System.IO;
-using System.Text;
 using System.Text.RegularExpressions;
 using System.Xml;
 
@@ -8,73 +6,8 @@ namespace WordAddIn1
 {
     public partial class RibbonMJS
     {
-        private (XmlDocument objXml, XmlDocument objToc, XmlDocument objBody) LoadAndProcessXml(string htmlStr, string docTitle)
-        {
-            XmlDocument objXml = new XmlDocument();
-            objXml.LoadXml(htmlStr);
-            ProcessXmlDocuments(objXml, docTitle, out XmlDocument objToc, out XmlDocument objBody);
-            return (objXml, objToc, objBody);
-        }
-
-        public void ProcessXmlDocuments(XmlDocument objXml, string docTitle, out XmlDocument objToc, out XmlDocument objBody)
-        {
-            // img ノードの height と width 属性を削除
-            foreach (XmlElement imgNode in objXml.SelectNodes("//img"))
-            {
-                imgNode.RemoveAttribute("height");
-                imgNode.RemoveAttribute("width");
-            }
-
-            // 不要なページ区切りを削除
-            foreach (XmlElement pageBreak in objXml.SelectNodes("//span[(translate(., ' &#10;&#13;&#9;', '') = '') and (count(*) = 1) and boolean(br[@style = 'page-break-before:always'])]"))
-            {
-                pageBreak.ParentNode.RemoveChild(pageBreak);
-            }
-            foreach (XmlElement pageBreak in objXml.SelectNodes("//br[translate(@style, ' &#10;&#13;&#9;', '') = 'page-break-before:always']"))
-            {
-                pageBreak.ParentNode.RemoveChild(pageBreak);
-            }
-
-            // コメントを削除
-            foreach (XmlElement comment in objXml.SelectNodes("//*[boolean(./*/@class[starts-with(., 'msocom')])]"))
-            {
-                comment.ParentNode.RemoveChild(comment);
-            }
-
-            // リンクのテキストを正規化
-            foreach (XmlElement link in objXml.SelectNodes("//a[boolean(@href)]"))
-            {
-                if (link.InnerText.Contains("http")) continue;
-                link.InnerText = Regex.Replace(link.InnerText, @"^(.*?)(?=[\s　](\d+\.\d+|[^\s|　]*?章))", "");
-                link.InnerText = Regex.Replace(link.InnerText, @"^[\s　]*(?:第[\d０-９]+章)*[\s　]+", "");
-                link.InnerText = Regex.Replace(link.InnerText, @"^[\s　]*(?:\d+\.)*\d+[\s　]+", "");
-            }
-
-            // 見出しで箇条書きタグを削除
-            foreach (XmlElement toc in objXml.SelectNodes("//a[starts-with(@name, '_Toc')]"))
-            {
-                foreach (XmlElement childSpan in toc.SelectNodes(".//span[contains(@style, 'Wingdings')]"))
-                    childSpan.ParentNode.RemoveChild(childSpan);
-                foreach (XmlElement brotherSpan in toc.ParentNode.SelectNodes(".//span[contains(@style, 'Wingdings')]"))
-                    brotherSpan.ParentNode.RemoveChild(brotherSpan);
-
-                // 目次テキスト先頭の「・」や「・　」を除去
-                if (!string.IsNullOrEmpty(toc.InnerText))
-                {
-                    toc.InnerText = Regex.Replace(toc.InnerText, @"^[・･]\s*", "");
-                }
-            }
-
-            // objToc と objBody の初期化
-            objToc = new XmlDocument();
-            objToc.LoadXml(@"<result><item title=""" + docTitle + @"""></item></result>");
-
-            objBody = new XmlDocument();
-            objBody.LoadXml("<result></result>");
-        }
-
-        // HTML セクションを解析し、目次 (objToc) と本文 (objBody) を構築する
-        public void GenerateTocAndBody(
+        // XML から目次 (objToc) と本文 (objBody) を構築する
+        public void BuildTocBodyFromXml(
             XmlDocument objXml,
             XmlDocument objBody,
             XmlDocument objToc,
@@ -327,6 +260,17 @@ namespace WordAddIn1
                     {
                         objBodyCurrent.AppendChild(objBody.CreateElement("p"));
                         ((XmlElement)objBodyCurrent.LastChild).SetAttribute("class", "Heading4");
+
+                        // 先頭の「・」や「・　」を除去
+                        if (childs.InnerText.StartsWith("・") || childs.InnerText.StartsWith("･"))
+                        {
+                            var firstTextNode = childs.SelectSingleNode(".//text()[1]");
+                            if (firstTextNode != null)
+                            {
+                                firstTextNode.InnerText = Regex.Replace(firstTextNode.InnerText, @"^[・･]\s*", "");
+                            }
+                        }
+
                         foreach (XmlNode childItem in childs.ChildNodes)
                         {
                             InnerNode(styleName, objBodyCurrent.LastChild, childItem);
@@ -336,6 +280,17 @@ namespace WordAddIn1
                     {
                         objBodyCurrent.AppendChild(objBody.CreateElement("p"));
                         ((XmlElement)objBodyCurrent.LastChild).SetAttribute("class", "Heading5");
+
+                        // 先頭の「・」や「･」を除去
+                        if (childs.InnerText.StartsWith("・") || childs.InnerText.StartsWith("･"))
+                        {
+                            var firstTextNode = childs.SelectSingleNode(".//text()[1]");
+                            if (firstTextNode != null)
+                            {
+                                firstTextNode.InnerText = Regex.Replace(firstTextNode.InnerText, @"^[・･]\s*", "");
+                            }
+                        }
+
                         foreach (XmlNode childItem in childs.ChildNodes)
                         {
                             InnerNode(styleName, objBodyCurrent.LastChild, childItem);
@@ -354,151 +309,5 @@ namespace WordAddIn1
                 if (breakFlg) break;
             }
         }
-
-        // 目次ファイルの作成
-        public void GenerateTocFiles(XmlDocument objToc, string rootPath, string exportDir, Dictionary<string, string> mergeScript)
-        {
-            string htmlToc = "";
-            string htmlToc1 = "";
-            string htmlToc2 = "";
-            string htmlToc3 = "";
-
-            foreach (XmlNode toc in objToc.SelectNodes("/result/item"))
-            {
-                htmlToc = @"{""type"":""book"",""name"":""" + ((XmlElement)toc).GetAttribute("title") + @""",""key"":""toc1""}";
-
-                foreach (XmlNode toc1 in toc.SelectNodes("item"))
-                {
-                    if (htmlToc1 != "")
-                    {
-                        htmlToc1 = htmlToc1 + ",";
-                    }
-
-                    htmlToc1 = htmlToc1 + @"{""type"":""" + (toc1.SelectNodes("item").Count != 0 ? "book" : "item") + @""",""name"":""" + ((XmlElement)toc1).GetAttribute("title") + @"""";
-
-                    if (toc1.SelectNodes("item").Count != 0)
-                    {
-                        htmlToc1 += @",""key"":""toc" + (toc1.SelectNodes("preceding::item[boolean(item)]").Count + 2) + @"""";
-                    }
-
-                    if (((XmlElement)toc1).GetAttribute("href") != "")
-                    {
-                        htmlToc1 += @",""url"":""" + makeHrefWithMerge(mergeScript, ((XmlElement)toc1).GetAttribute("href")) + @"""";
-                    }
-
-                    htmlToc1 += "}";
-
-                    foreach (XmlNode toc2 in toc1.SelectNodes("item"))
-                    {
-                        if (htmlToc2 != "")
-                        {
-                            htmlToc2 = htmlToc2 + ",";
-                        }
-
-                        htmlToc2 += @"{""type"":""" + (toc2.SelectNodes("item").Count != 0 ? "book" : "item") + @""",""name"":""" + ((XmlElement)toc2).GetAttribute("title") + @"""";
-
-                        if (toc2.SelectNodes("item").Count != 0)
-                        {
-                            htmlToc2 += @",""key"":""toc" + (toc2.SelectNodes("preceding::item[boolean(item)]").Count + 3) + @"""";
-                        }
-                        if (((XmlElement)toc2).GetAttribute("href") != "")
-                        {
-                            htmlToc2 += @",""url"":""" + makeHrefWithMerge(mergeScript, ((XmlElement)toc2).GetAttribute("href")) + @"""";
-                        }
-
-                        htmlToc2 += "}";
-
-                        foreach (XmlNode toc3 in toc2.SelectNodes("item"))
-                        {
-                            if (htmlToc3 != "")
-                            {
-                                htmlToc3 += ",";
-                            }
-
-                            htmlToc3 += @"{""type"":""item"",""name"":""" + ((XmlElement)toc3).GetAttribute("title") + @""",""url"":""" + makeHrefWithMerge(mergeScript, ((XmlElement)toc3).GetAttribute("href")) + @"""}";
-                        }
-
-                        if (htmlToc3 != "")
-                        {
-                            using (StreamWriter sw = new StreamWriter(rootPath + "\\" + exportDir + "\\whxdata\\toc" + (toc2.SelectNodes("preceding::item[boolean(item)]").Count + 3) + ".new.js", false, Encoding.UTF8))
-                            {
-                                sw.WriteLine("(function() {");
-                                sw.WriteLine("var toc =  [" + htmlToc3 + "];");
-                                sw.WriteLine("window.rh.model.publish(rh.consts('KEY_TEMP_DATA'), toc, { sync:true });");
-                                sw.WriteLine("})();");
-                            }
-                            htmlToc3 = "";
-                        }
-                    }
-
-                    if (htmlToc2 != "")
-                    {
-                        using (StreamWriter sw = new StreamWriter(rootPath + "\\" + exportDir + "\\whxdata\\toc" + (toc1.SelectNodes("preceding::item[boolean(item)]").Count + 2) + ".new.js", false, Encoding.UTF8))
-                        {
-                            sw.WriteLine("(function() {");
-                            sw.WriteLine("var toc =  [" + htmlToc2 + "];");
-                            sw.WriteLine("window.rh.model.publish(rh.consts('KEY_TEMP_DATA'), toc, { sync:true });");
-                            sw.WriteLine("})();");
-                        }
-                        htmlToc2 = "";
-                    }
-                }
-
-                using (StreamWriter sw = new StreamWriter(rootPath + "\\" + exportDir + "\\whxdata\\toc1.new.js", false, Encoding.UTF8))
-                {
-                    sw.WriteLine("(function() {");
-                    sw.WriteLine("var toc =  [" + htmlToc1 + "];");
-                    sw.WriteLine("window.rh.model.publish(rh.consts('KEY_TEMP_DATA'), toc, { sync:true });");
-                    sw.WriteLine("})();");
-                }
-            }
-
-            using (StreamWriter sw = new StreamWriter(rootPath + "\\" + exportDir + "\\whxdata\\toc.new.js", false, Encoding.UTF8))
-            {
-                sw.WriteLine("(function() {");
-                sw.WriteLine("var toc =  [" + htmlToc + "];");
-                sw.WriteLine("window.rh.model.publish(rh.consts('KEY_TEMP_DATA'), toc, { sync:true });");
-                sw.WriteLine("})();");
-            }
-        }
-
-        public void CleanUpXmlNodes(XmlDocument objBody)
-        {
-            // lang 属性や name 属性を削除し、不要なノードを整理
-            foreach (XmlElement langSpan in objBody.SelectNodes(".//span[boolean(@lang)]|.//a"))
-            {
-                langSpan.RemoveAttribute("lang");
-
-                if (langSpan.Name == "a")
-                {
-                    langSpan.RemoveAttribute("name");
-                }
-
-                if (langSpan.Attributes.Count == 0)
-                {
-                    while (langSpan.ChildNodes.Count != 0)
-                    {
-                        langSpan.ParentNode.InsertBefore(langSpan.ChildNodes[0], langSpan);
-                    }
-                    langSpan.ParentNode.RemoveChild(langSpan);
-                }
-            }
-
-            // 不要な <div> や <br> タグを削除
-            while (objBody.SelectSingleNode("/result/div//*[((name() = 'div') or (name() = 'br')) and not(boolean(ancestor::table)) and not(boolean(node()[translate(., ' &#9;" + ((char)160).ToString() + "　', '') != ''])) and not(boolean(following-sibling::node()[translate(., ' &#9;" + ((char)160).ToString() + "　', '') != '']))] ") != null)
-            {
-                XmlNode lineBreak = objBody.SelectSingleNode("/result/div//*[((name() = 'div') or (name() = 'br')) and not(boolean(ancestor::table)) and not(boolean(node()[translate(., ' &#9;" + ((char)160).ToString() + "　', '') != ''])) and not(boolean(following-sibling::node()[translate(., ' &#9;" + ((char)160).ToString() + "　', '') != '']))] ");
-                lineBreak.ParentNode.RemoveChild(lineBreak);
-            }
-
-            // 不要な <p> タグを削除
-            while (objBody.SelectSingleNode("/result/div//*[not(img)][(name() = 'p') and not(boolean(ancestor::table)) and not(boolean(node()[translate(., ' &#9;" + ((char)160).ToString() + "　', '') != ''])) and not(boolean(following-sibling::node()[translate(., ' &#9;" + ((char)160).ToString() + "　', '') != '']))] ") != null)
-            {
-                XmlNode lineBreak = objBody.SelectSingleNode("/result/div//*[not(img)][(name() = 'p') and not(boolean(ancestor::table)) and not(boolean(node()[translate(., ' &#9;" + ((char)160).ToString() + "　', '') != ''])) and not(boolean(following-sibling::node()[translate(., ' &#9;" + ((char)160).ToString() + "　', '') != '']))] ");
-                lineBreak.ParentNode.RemoveChild(lineBreak);
-            }
-
-        }
-
     }
 }
