+++ src/MJS_WordPlugin_src/WordAddIn1/RibbonMJS.InnerNode.cs
@@ -152,35 +152,44 @@ namespace WordAddIn1
                         AppendChoiceElement(objTargetNode, seekNode, thisStyleName);
                     }
 
+                    //else if (Regex.IsMatch(thisStyleName, "箇条書き[2２]"))
+                    //{
+                    //    objTargetNode.AppendChild(objTargetNode.OwnerDocument.CreateElement("p"));
+                    //    ((XmlElement)objTargetNode.LastChild).RemoveAttribute("style");
+                    //    if (Regex.Replace(seekNode.InnerText, @"[\s　\u00A0]", "") == "")
+                    //    {
+                    //        ((XmlElement)objTargetNode.LastChild).RemoveAttribute("class");
+                    //    }
+
+                    //    if ((Regex.Replace(seekNode.InnerText, @"[\s　\u00A0]", "") != "") && (objTargetNode.SelectNodes("ancestor-or-self::*[@class = 'MJS_listItem2']").Count == 0))
+                    //    {
+                    //        seekNode.InnerText = Regex.Replace(seekNode.InnerText, @"^\S{0,3}[ 　]+", "");
+                    //        ((XmlElement)objTargetNode.LastChild).SetAttribute("class", "MJS_listItem2");
+                    //    }
+                    //}
+                    //else if (Regex.IsMatch(thisStyleName, "箇条書き"))
+                    //{
+                    //    objTargetNode.AppendChild(objTargetNode.OwnerDocument.CreateElement("p"));
+                    //    ((XmlElement)objTargetNode.LastChild).RemoveAttribute("style");
+                    //    if (Regex.Replace(seekNode.InnerText, @"[\s　\u00A0]", "") == "")
+                    //    {
+                    //        ((XmlElement)objTargetNode.LastChild).RemoveAttribute("class");
+                    //    }
+
+                    //    if ((Regex.Replace(seekNode.InnerText, @"[\s　\u00A0]", "") != "") && (objTargetNode.SelectNodes("ancestor-or-self::*[@class = 'MJS_listItem']").Count == 0))
+                    //    {
+                    //        seekNode.InnerText = Regex.Replace(seekNode.InnerText, @"^\S{0,3}[ 　]+", "");
+                    //        ((XmlElement)objTargetNode.LastChild).SetAttribute("class", "MJS_listItem");
+                    //    }
+                    //}
+
                     else if (Regex.IsMatch(thisStyleName, "箇条書き[2２]"))
                     {
-                        objTargetNode.AppendChild(objTargetNode.OwnerDocument.CreateElement("p"));
-                        ((XmlElement)objTargetNode.LastChild).RemoveAttribute("style");
-                        if (Regex.Replace(seekNode.InnerText, @"[\s　\u00A0]", "") == "")
-                        {
-                            ((XmlElement)objTargetNode.LastChild).RemoveAttribute("class");
-                        }
-
-                        if ((Regex.Replace(seekNode.InnerText, @"[\s　\u00A0]", "") != "") && (objTargetNode.SelectNodes("ancestor-or-self::*[@class = 'MJS_listItem2']").Count == 0))
-                        {
-                            seekNode.InnerText = Regex.Replace(seekNode.InnerText, @"^\S{0,3}[ 　]+", "");
-                            ((XmlElement)objTargetNode.LastChild).SetAttribute("class", "MJS_listItem2");
-                        }
+                        AppendListItemElement(objTargetNode, seekNode, "MJS_listItem2");
                     }
                     else if (Regex.IsMatch(thisStyleName, "箇条書き"))
                     {
-                        objTargetNode.AppendChild(objTargetNode.OwnerDocument.CreateElement("p"));
-                        ((XmlElement)objTargetNode.LastChild).RemoveAttribute("style");
-                        if (Regex.Replace(seekNode.InnerText, @"[\s　\u00A0]", "") == "")
-                        {
-                            ((XmlElement)objTargetNode.LastChild).RemoveAttribute("class");
-                        }
-
-                        if ((Regex.Replace(seekNode.InnerText, @"[\s　\u00A0]", "") != "") && (objTargetNode.SelectNodes("ancestor-or-self::*[@class = 'MJS_listItem']").Count == 0))
-                        {
-                            seekNode.InnerText = Regex.Replace(seekNode.InnerText, @"^\S{0,3}[ 　]+", "");
-                            ((XmlElement)objTargetNode.LastChild).SetAttribute("class", "MJS_listItem");
-                        }
+                        AppendListItemElement(objTargetNode, seekNode, "MJS_listItem");
                     }
 
                     else if (Regex.IsMatch(thisStyleName, "表内-項目_センタリング"))
